Deprecate DepthMaterial
