+++ src/deprecated/deprecated.js
@@ -55,7 +55,6 @@ import { PROJECTION_ORTHOGRAPHIC, PROJECTION_PERSPECTIVE, LAYERID_IMMEDIATE, LIN
 import { calculateTangents, createBox, createCapsule, createCone, createCylinder, createMesh, createPlane, createSphere, createTorus } from '../scene/procedural.js';
 import { partitionSkin } from '../scene/skin-partition.js';
 import { BasicMaterial } from '../scene/materials/basic-material.js';
-import { DepthMaterial } from '../scene/materials/depth-material.js';
 import { ForwardRenderer } from '../scene/renderer/forward-renderer.js';
 import { GraphNode } from '../scene/graph-node.js';
 import { Material } from '../scene/materials/material.js';
@@ -464,6 +463,19 @@ Object.defineProperties(Texture.prototype, {
 });
 
 // SCENE
+
+class DepthMaterial extends Material {
+    updateShader(device) {
+        Debug.deprecated("pc.DepthMaterial is deprecated.");
+        const options = {
+            skin: !!this.meshInstances[0].skinInstance
+        };
+        const library = device.getProgramLibrary();
+        this.shader = library.getProgram('depth', options);
+    }
+}
+
+export { DepthMaterial };
 
 export const PhongMaterial = StandardMaterial;
